TES-42 refactor: folder structure for UI
