+++ __test__/Home.test.jsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
-import Home from '../app/page';
+import Home from '../src/app/page';
 
 describe('Page', () => {
   it('renders a heading', () => {
