+++ __test__/snapshot.js
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import Page from '../app/page';
+import Page from '../src/app/page';
 
 it('renders homepage unchanged', () => {
   const { container } = render(<Page />);
