--- app/page.tsx
@@ -1,12 +1,0 @@
-import Link from 'next/link';
-
-export default function Home() {
-  return (
-    <main className='flex min-h-screen flex-col items-center justify-between p-24'>
-      <div>
-        <h1>Home</h1>
-        <Link href='/about'>About</Link>
-      </div>
-    </main>
-  );
-}
