+++ package.json
@@ -14,7 +14,8 @@
   "dependencies": {
     "next": "14.2.5",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "telia-e-shop-frontend": "file:"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.4.8",
