+++ src/components/TestButton.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const TestButton = () => {
+  return (
+    <button className='rounded-md bg-yellow-600 font-bold text-white px-4 py-2'>
+      test button
+    </button>
+  );
+};
